requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
Flask-MySQLdb==1.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
gevent>=1.4
hvac==0.11.2
mysql-connector==2.2.9
PyJWT==2.4.0
PyYAML==5.3.1
requests==2.27.1